remove endpoint from SDK configs, add UI for empty lobby list, made docs header linkable, better organize react components
