+++ client/src/components/lobby/Header.tsx
