+++ client/src/components/website/HathoraLogo.tsx
