+++ client/src/input.css
@@ -2,6 +2,8 @@
 @tailwind components;
 @tailwind utilities;
 
+html{scroll-behavior:smooth}
+
 .loading-dots-animation {
     display:inline-block;
     clip-path: inset(0 1ch 0 0);
